requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
mysqlclient==2.1.1
pydantic[email]